*mod* model edited ingredient.java quant to be float (from interceptor error msg) *upd* updated retroservice to contain static instance of OkHttpClient (to which interceptor added) *man* added internet permission

+++ app/src/main/java/com/aaqanddev/aaqsbakingapp/model/Ingredient.java
@@ -12,7 +12,7 @@ public class Ingredient {
 
     @SerializedName("quantity")
     @Expose
-    private Integer quantity;
+    private float quantity;
     @SerializedName("measure")
     @Expose
     private String measure;
@@ -20,11 +20,11 @@ public class Ingredient {
     @Expose
     private String ingredient;
 
-    public Integer getQuantity() {
+    public float getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(float quantity) {
         this.quantity = quantity;
     }
